# internal/server/docs/swagger.yaml
components:
  schemas:
...
    api.APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        message:
          type: string
      type: object
...
    api.AddPropertyResponse:
      properties:
        propId:
          type: string
      type: object
...
    api.ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/api.APIError'
      type: object
...
    api.PropertiesListResponse:
      properties:
        properties:
          items:
            $ref: '#/components/schemas/api.PropertyDef'
          type: array
          uniqueItems: false
      type: object
...
    api.PropertyPatchRequest:
      properties:
        set:
          type: object
        unset:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
...
    api.TypeInfo:
      properties:
        builtIn:
          type: boolean
        description:
          type: string
        hidden:
          description: |-
            Hidden / Meta — see TypesCreateRequest. GET …/types omits hidden
            types unless includeHidden=true.
          type: boolean
        iconCid:
          type: string
        id:
          type: string
        layout:
          additionalProperties: {}
          description: Layout — see TypesCreateRequest. Absent on built-ins.
          type: object
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
        xKey:
          description: |-
            XKey is the stable, caller-side programmatic key. For builtin/registered
            types it equals Id (a clean literal like "chat"); for user types it's
            the value set at create (or derived from Name by the client). Clients use
            it as the stable type handle in dotted property paths.
          type: string
      type: object
...
    api.TypePatchRequest:
      properties:
        description:
          type: string
        hidden:
          type: boolean
        iconCid:
          type: string
        layout:
          additionalProperties: {}
          type: object
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
      type: object
    api.TypesCreateRequest:
      properties:
        description:
          type: string
        hidden:
          description: |-
            Hidden keeps the type out of GET …/types by default (pass
            includeHidden=true to list it) and out of a client's pickers;
            GET …/types/:typeId always resolves it. Meta is the open bag of
            consumer flags — one string, bool or number per single-level key
            (no '.', no '$', ≤64 bytes), written per key so concurrent
            writers merge; opaque to the server. Both mutable via PATCH.
          type: boolean
        iconCid:
          type: string
        layout:
          additionalProperties: {}
          description: |-
            Layout is how the type's header and parts compose — {type,
            config} in the xFormat shape (v1 slugs: page, tabs, chat, profile;
            open set, unknown renders as page). Mutable via PATCH
            …/types/:typeId.
          type: object
        meta:
          additionalProperties: {}
          type: object
        name:
          type: string
        xKey:
          description: |-
            XKey is the stable, caller-side programmatic key for the type
            (e.g. "agent_memory"). REQUIRED on create and unique per space: the
            server rejects an empty xKey (type.xkey_required) and one that
            collides with an existing type's xKey or id (type.xkey_conflict).
            Clients derive it as a slug of Name. It's the only human handle a
            type resolves by — the display Name is not a resolution key.
          type: string
      type: object
    api.TypesCreateResponse:
      properties:
        typeId:
          type: string
      type: object
    api.TypesListResponse:
      properties:
        types:
          items:
            $ref: '#/components/schemas/api.TypeInfo'
          type: array
          uniqueItems: false
      type: object
...
info:
  description: Local HTTP/JSON API wrapping any-sync-sdk. Localhost-only, no auth
    in v1.
  title: Any API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceId}/types:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.TypesListResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List types in a space
      tags:
      - types
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.TypesCreateRequest'
                description: Type params
                summary: body
        description: Type params
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.TypesCreateResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Create a type
      tags:
      - types
  /spaces/{spaceId}/types/{typeId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.TypeInfo'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get a type
      tags:
      - types
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.TypePatchRequest'
                description: Fields to change
                summary: body
        description: Fields to change
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a type's display and rendering metadata
      tags:
      - types
...
  /spaces/{spaceId}/types/{typeId}/properties:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PropertiesListResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: List properties of a type
      tags:
      - types
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.AddPropertyRequest'
                description: Property draft
                summary: body
        description: Property draft
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.AddPropertyResponse'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Add a property to a type
      tags:
      - types
  /spaces/{spaceId}/types/{typeId}/properties/{propId}:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Remove a property definition
      tags:
      - types
    patch:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      - description: Property ID
        in: path
        name: propId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.PropertyPatchRequest'
                description: set/unset paths
                summary: body
        description: set/unset paths
        required: true
      responses:
        "204":
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Patch a property definition (rename, options, colors, order)
      tags:
      - types
...
servers:
- url: http://127.0.0.1:7001/v1