# internal/server/handlers_types.go
package server

import (
	"context"
	"encoding/json"
	"errors"
	"net/http"

	"github.com/labstack/echo/v4"

	"github.com/anyproto/any-sync-sdk/space"

	"github.com/anyproto/any/internal/api"
	"github.com/anyproto/any/internal/index"
)

// typeCreate handles POST /v1/spaces/:spaceId/types.
//
//	@Summary	Create a type
//	@Tags		types
//	@Accept		json
//	@Produce	json
//	@Param		spaceId	path		string					true	"Space ID"
//	@Param		body	body		api.TypesCreateRequest	true	"Type params"
//	@Success	201		{object}	api.TypesCreateResponse
//	@Failure	400		{object}	api.ErrorEnvelope
//	@Failure	409		{object}	api.ErrorEnvelope
//	@Failure	500		{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types [post]
func (d *deps) typeCreate(c echo.Context) error {
	// Strict bind: type create carries NO inline property definitions —
	// no SDK surface accepts them. A lax bind would drop a `properties`
	// array silently and read as "properties never sync" (the schema
	// stays empty forever), so unknown keys reject loudly with the
	// per-field route named in the hint.
	req, ok := bindBodyStrict[api.TypesCreateRequest](c,
		"inline property definitions are not part of type create — create the type, "+
			"then add each property via POST /v1/spaces/{spaceId}/types/{typeId}/properties")
	if !ok {
		return nil
	}

	// xKey is the only human-supplied handle a type can be resolved by
	// (besides its CID) — the display name is NOT a resolution key. The SDK
	// treats xKey as non-unique display metadata, so uniqueness is enforced
	// here. Reject name-only creates and same-space collisions; otherwise a
	// type is reachable only by its content-addressed id (see docs/03-api.md
	// § Types).
	if req.XKey == "" {
		return writeError(c, http.StatusBadRequest, "type.xkey_required",
			"xKey is required: a type needs a stable programmatic handle (derive a slug from the name)", nil)
	}

	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}

	if errResp, done := requireHandleFree(c, sp, req.XKey); done {
		return errResp
	}

	layout, code, reason := layoutFromWire(req.Layout)
	if code != "" {
		return writeError(c, http.StatusBadRequest, code, reason, map[string]any{"path": "layout"})
	}
	for k, v := range req.Meta {
		if code, reason := checkTypeMetaEntry(k, v); code != "" {
			return writeError(c, http.StatusBadRequest, code, reason, map[string]any{"path": "meta." + k})
		}
	}
	typeId, err := sp.Types().Create(c.Request().Context(), space.TypeCreateParams{
		Name:        req.Name,
		Description: req.Description,
		IconCID:     req.IconCID,
		XKey:        req.XKey,
		Layout:      layout,
		Hidden:      req.Hidden,
		Meta:        req.Meta,
	})
	if err != nil {
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id()})
	}
	return c.JSON(http.StatusCreated, api.TypesCreateResponse{TypeId: typeId})
}

// typeAddProperty handles POST /v1/spaces/:spaceId/types/:typeId/properties.
//
//	@Summary	Add a property to a type
//	@Tags		types
//	@Accept		json
//	@Produce	json
//	@Param		spaceId	path		string					true	"Space ID"
//	@Param		typeId	path		string					true	"Type ID"
//	@Param		body	body		api.AddPropertyRequest	true	"Property draft"
//	@Success	201		{object}	api.AddPropertyResponse
//	@Failure	400		{object}	api.ErrorEnvelope
//	@Failure	500		{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types/{typeId}/properties [post]
func (d *deps) typeAddProperty(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	typeId := c.Param("typeId")
	if typeId == "" {
		return writeError(c, http.StatusBadRequest, "request.missing_field", "typeId required", nil)
	}

	req, ok := bindBodyStrict[api.AddPropertyRequest](c, "")
	if !ok {
		return nil
	}
	draft, code, reason, details := propertyDraftFromAPI(*req)
	if code != "" {
		return writeError(c, http.StatusBadRequest, code, reason, details)
	}

	// xKey is unique within the type — a read-then-create preflight,
	// not a guarantee (two devices working apart can both land the same
	// handle; both columns then persist — docs/27-descriptors.md).
	if req.XKey != "" {
		defs, err := sp.Types().Properties(c.Request().Context(), typeId)
		if err != nil {
			return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
		}
		if errResp, done := requireXKeyFree(c, defs, req.XKey, ""); done {
			return errResp
		}
	}

	propId, err := sp.Types().AddProperty(c.Request().Context(), typeId, draft)
	if err != nil {
		if errors.Is(err, space.ErrTypeRegistered) {
			return writeError(c, http.StatusBadRequest, "type.registered",
				"type is a registered built-in; its properties are statically declared",
				map[string]any{"spaceId": sp.Id(), "typeId": typeId})
		}
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
	}
	return c.JSON(http.StatusCreated, api.AddPropertyResponse{PropId: propId})
}

// propertyDraftFromAPI is the one gate a property definition passes on
// its way to the SDK — POST …/properties and a bundle's `properties`
// alike: kind required and known (nothing is defaulted from the
// descriptor), meta narrowed to the index flag, the descriptor
// validated against the kind, scope creatable. Returns ("", "", nil)
// code/reason/details on success.
func propertyDraftFromAPI(req api.AddPropertyRequest) (space.PropertyDraft, string, string, map[string]any) {
	var draft space.PropertyDraft
	if req.Kind == "" {
		return draft, "request.schema", "kind is required (string / number / boolean / array / object / datetime)", nil
	}
	kind, ok := propertyKindFromString(req.Kind)
	if !ok {
		return draft, "request.schema", "unknown property kind", map[string]any{"kind": req.Kind}
	}
	for k := range req.Meta {
		if k != index.MetaIndexKey {
			return draft, "request.invalid_field",
				"meta holds only " + index.MetaIndexKey + "; descriptive keys live under xFormat",
				map[string]any{"key": k}
		}
	}
	xf, code, reason := validateDescriptor(req.XFormat, req.Kind)
	if code != "" {
		return draft, code, reason, nil
	}
	var scope space.Scope // zero value = synced (SDK default)
	if req.Scope != "" {
		scope, ok = space.ParseScope(req.Scope)
		if !ok || scope == space.ScopeDerived {
			return draft, "request.schema", "scope must be one of synced, account, local", map[string]any{"scope": req.Scope}
		}
	}
	return space.PropertyDraft{
		Name:        req.Name,
		Description: req.Description,
		XKey:        req.XKey,
		Kind:        kind,
		Meta:        req.Meta,
		XFormat:     xf,
		Scope:       scope,
	}, "", "", nil
}

// requireXKeyFree 409s when another of the type's definitions already
// carries xKey (selfId excludes the property being patched). An unknown
// type lists no properties, so the write that follows reports it.
func requireXKeyFree(c echo.Context, defs []space.PropertyDef, xKey, selfId string) (errResp error, done bool) {
	for _, def := range defs {
		if def.XKey == xKey && def.Id != selfId {
			return writeError(c, http.StatusConflict, "property.xkey_conflict",
				"xKey already in use by another property of this type",
				map[string]any{"xKey": xKey, "existingPropId": def.Id}), true
		}
	}
	return nil, false
}

// typeList handles GET /v1/spaces/:spaceId/types.
//
//	@Summary	List types in a space
//	@Tags		types
//	@Produce	json
//	@Param		spaceId	path		string	true	"Space ID"
//	@Success	200		{object}	api.TypesListResponse
//	@Failure	500		{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types [get]
func (d *deps) typeList(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	infos, err := sp.Types().List(c.Request().Context())
	if err != nil {
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id()})
	}
	// Hidden types (a client's choice, a bundle install's `hidden`, the
	// built-in page/miniapp/bin/dataview) stay out of the default listing
	// — the picker view — and come back with includeHidden=true;
	// GET …/types/:typeId resolves them always.
	includeHidden := c.QueryParam("includeHidden") == "true"
	out := make([]api.TypeInfo, 0, len(infos))
	for _, t := range infos {
		if t.Hidden && !includeHidden {
			continue
		}
		out = append(out, typeInfoToAPI(t))
	}
	return c.JSON(http.StatusOK, api.TypesListResponse{Types: out})
}

// typeGet handles GET /v1/spaces/:spaceId/types/:typeId.
//
//	@Summary	Get a type
//	@Tags		types
//	@Produce	json
//	@Param		spaceId	path		string	true	"Space ID"
//	@Param		typeId	path		string	true	"Type ID"
//	@Success	200		{object}	api.TypeInfo
//	@Failure	404		{object}	api.ErrorEnvelope
//	@Failure	500		{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types/{typeId} [get]
func (d *deps) typeGet(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	typeId := c.Param("typeId")
	if typeId == "" {
		return writeError(c, http.StatusBadRequest, "request.missing_field", "typeId required", nil)
	}
	info, err := sp.Types().Get(c.Request().Context(), typeId)
	if err != nil {
		if resp, done := typeLookupError(c, err, sp.Id(), typeId); done {
			return resp
		}
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
	}
	return c.JSON(http.StatusOK, typeInfoToAPI(info))
}

// typeProperties handles GET /v1/spaces/:spaceId/types/:typeId/properties.
//
//	@Summary	List properties of a type
//	@Tags		types
//	@Produce	json
//	@Param		spaceId	path		string	true	"Space ID"
//	@Param		typeId	path		string	true	"Type ID"
//	@Success	200		{object}	api.PropertiesListResponse
//	@Failure	404		{object}	api.ErrorEnvelope
//	@Failure	500		{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types/{typeId}/properties [get]
func (d *deps) typeProperties(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	typeId := c.Param("typeId")
	if typeId == "" {
		return writeError(c, http.StatusBadRequest, "request.missing_field", "typeId required", nil)
	}
	// The SDK's Properties returns an empty slice for an unknown typeId
	// (the type's defs collection simply doesn't exist), which is
	// indistinguishable from "type exists, no properties yet". Check
	// existence first so a bad id is a typed 404, not a silent 200 [].
	if _, err := sp.Types().Get(c.Request().Context(), typeId); err != nil {
		if resp, done := typeLookupError(c, err, sp.Id(), typeId); done {
			return resp
		}
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
	}
	return d.propertiesOf(c, sp, typeId)
}

// typeLookupError maps the two typed misses of a type read: 404 for an
// unknown id, 400 type.not_a_type when the id names a collection.
func typeLookupError(c echo.Context, err error, spaceId, typeId string) (error, bool) {
	details := map[string]any{"spaceId": spaceId, "typeId": typeId}
	switch {
	case errors.Is(err, space.ErrNotAType):
		return writeError(c, http.StatusBadRequest, "type.not_a_type",
			"the id names a collection — use the …/collections routes", details), true
	case errors.Is(err, space.ErrNotFound):
		return writeError(c, http.StatusNotFound, "type.not_found", "type not found", details), true
	}
	return nil, false
}

// propertiesOf renders the definitions of a type or collection the
// caller has already resolved.
func (d *deps) propertiesOf(c echo.Context, sp space.Space, ownerId string) error {
	typeId := ownerId
	defs, err := sp.Types().Properties(c.Request().Context(), typeId)
	if err != nil {
		if errors.Is(err, space.ErrNotFound) {
			return writeError(c, http.StatusNotFound, "type.not_found",
				"type not found",
				map[string]any{"spaceId": sp.Id(), "typeId": typeId})
		}
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
	}
	out := make([]api.PropertyDef, 0, len(defs))
	for _, p := range defs {
		out = append(out, propertyDefToAPI(p))
	}
	return c.JSON(http.StatusOK, api.PropertiesListResponse{Properties: out})
}

// typePatchProperty handles PATCH /v1/spaces/:spaceId/types/:typeId/properties/:propId.
// It wraps TypesAPI.PatchProperty — a generic per-path patch covering
// rename, the handle, the index flag and every descriptor path (slug,
// icon, order, options, relation, config). Body:
// {set: {"dotted.path": value}, unset: ["dotted.path"]}. Pinned paths
// (kind/scope/items/properties) return 400 property.immutable.
//
//	@Summary	Patch a property definition (rename, options, colors, order)
//	@Tags		types
//	@Accept		json
//	@Param		spaceId	path	string						true	"Space ID"
//	@Param		typeId	path	string						true	"Type ID"
//	@Param		propId	path	string						true	"Property ID"
//	@Param		body	body	api.PropertyPatchRequest	true	"set/unset paths"
//	@Success	204
//	@Failure	400	{object}	api.ErrorEnvelope
//	@Failure	404	{object}	api.ErrorEnvelope
//	@Failure	500	{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types/{typeId}/properties/{propId} [patch]
func (d *deps) typePatchProperty(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	typeId := c.Param("typeId")
	propId := c.Param("propId")
	if typeId == "" || propId == "" {
		return writeError(c, http.StatusBadRequest, "request.missing_field", "typeId and propId required", nil)
	}

	req, ok := bindBodyStrict[api.PropertyPatchRequest](c, "")
	if !ok {
		return nil
	}
	if len(req.Set) == 0 && len(req.Unset) == 0 {
		return writeError(c, http.StatusBadRequest, "request.missing_field",
			"at least one of set/unset is required", nil)
	}

	patch := space.PropertyPatch{}
	if len(req.Set) > 0 {
		patch.Set = make(map[string]any, len(req.Set))
	}
	var newXKey, newSlug, newMode string
	for path, raw := range req.Set {
		storagePath, code, reason := patchPathToStorage(path, true)
		if code != "" {
			return writeError(c, http.StatusBadRequest, code, reason, map[string]any{"path": path})
		}
		val, vcode, reason := patchSetValue(storagePath, raw)
		if vcode != "" {
			return writeError(c, http.StatusBadRequest, vcode, reason, map[string]any{"path": path})
		}
		patch.Set[storagePath] = val
		switch storagePath {
		case propFieldXKey:
			newXKey, _ = val.(string)
		case propFieldXFormat + "." + xfType:
			newSlug, _ = val.(string)
		case propFieldXFormat + "." + xfLinks:
			newMode, _ = val.(string)
		}
	}
	for _, path := range req.Unset {
		storagePath, code, reason := patchPathToStorage(path, false)
		if code != "" {
			return writeError(c, http.StatusBadRequest, code, reason, map[string]any{"path": path})
		}
		patch.Unset = append(patch.Unset, storagePath)
	}

	// The two leaves with a cross-definition rule: a handle must stay
	// unique within the type, and a slug can only move within the
	// pinned kind. Both need the current definitions — one read, only
	// when either leaf is touched.
	if newXKey != "" || newSlug != "" || newMode != "" {
		defs, err := sp.Types().Properties(c.Request().Context(), typeId)
		if err != nil {
			return sdkOpError(c, err, map[string]any{"spaceId": sp.Id(), "typeId": typeId})
		}
		if newXKey != "" {
			if errResp, done := requireXKeyFree(c, defs, newXKey, propId); done {
				return errResp
			}
		}
		for _, def := range defs {
			if newSlug == "" || def.Id != propId {
				continue
			}
			if reason := slugKindMismatch(newSlug, propertyKindToString(def.Kind)); reason != "" {
				return writeError(c, http.StatusBadRequest, "property.format_invalid", reason,
					map[string]any{"path": wireXFormat + "." + xfType})
			}
		}
		if newMode != "" {
			for _, def := range defs {
				if def.Id != propId {
					continue
				}
				if reason := linkModeMismatch(newMode, propertyKindToString(def.Kind)); reason != "" {
					return writeError(c, http.StatusBadRequest, "property.format_invalid", reason,
						map[string]any{"path": wireXFormat + "." + xfLinks})
				}
			}
		}
	}

	if err := sp.Types().PatchProperty(c.Request().Context(), typeId, propId, patch); err != nil {
		return d.propertyWriteError(c, err, typeId, propId)
	}
	return c.NoContent(http.StatusNoContent)
}

// propertyWriteError maps the shared client-error sentinels from
// PatchProperty / RemoveProperty onto clean 4xx envelopes (never a 500
// leaking the SDK's internal "typesAPI:" message).
func (d *deps) propertyWriteError(c echo.Context, err error, typeId, propId string) error {
	details := map[string]any{"typeId": typeId, "propId": propId}
	switch {
	case errors.Is(err, space.ErrNotFound):
		return writeError(c, http.StatusNotFound, "sdk.not_found", "type or property not found", details)
	case errors.Is(err, space.ErrPinnedField):
		return writeError(c, http.StatusBadRequest, "property.immutable", "a patched path is immutable", details)
	case errors.Is(err, space.ErrTypeRegistered):
		return writeError(c, http.StatusBadRequest, "type.registered",
			"type is a registered built-in; its properties are statically declared", details)
	default:
		return sdkOpError(c, err, details)
	}
}

// typeRemoveProperty handles DELETE /v1/spaces/:spaceId/types/:typeId/properties/:propId.
// It wraps TypesAPI.RemoveProperty — a synced tombstone of the property
// definition. Existing instance values are not cleaned up (dangling-
// tolerant). Unknown/already-removed propId → 404.
//
//	@Summary	Remove a property definition
//	@Tags		types
//	@Param		spaceId	path	string	true	"Space ID"
//	@Param		typeId	path	string	true	"Type ID"
//	@Param		propId	path	string	true	"Property ID"
//	@Success	204
//	@Failure	404	{object}	api.ErrorEnvelope
//	@Failure	500	{object}	api.ErrorEnvelope
//	@Router		/spaces/{spaceId}/types/{typeId}/properties/{propId} [delete]
func (d *deps) typeRemoveProperty(c echo.Context) error {
	sp, errResp, done := d.resolveSpace(c)
	if done {
		return errResp
	}
	typeId := c.Param("typeId")
	propId := c.Param("propId")
	if typeId == "" || propId == "" {
		return writeError(c, http.StatusBadRequest, "request.missing_field", "typeId and propId required", nil)
	}
	if err := sp.Types().RemoveProperty(c.Request().Context(), typeId, propId); err != nil {
		return d.propertyWriteError(c, err, typeId, propId)
	}
	return c.NoContent(http.StatusNoContent)
}

// requireHandleFree refuses an xKey any type or collection of the
// space already resolves by — registered definitions resolve by their
// literal Id ("page", "miniapp", …) as well as by xKey, so a new xKey
// must dodge both namespaces, on both surfaces.
func requireHandleFree(c echo.Context, sp space.Space, xKey string) (errResp error, done bool) {
	ctx := c.Request().Context()
	types, err := sp.Types().List(ctx)
	if err != nil {
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id()}), true
	}
	for _, t := range types {
		if t.XKey == xKey || t.Id == xKey {
			return writeError(c, http.StatusConflict, "type.xkey_conflict",
				"xKey already in use by another type in this space",
				map[string]any{"xKey": xKey, "existingTypeId": t.Id}), true
		}
	}
	colls, err := sp.Collections().List(ctx)
	if err != nil {
		return sdkOpError(c, err, map[string]any{"spaceId": sp.Id()}), true
	}
	for _, col := range colls {
		if col.XKey == xKey || col.Id == xKey {
			return writeError(c, http.StatusConflict, "type.xkey_conflict",
				"xKey already in use by a collection in this space",
				map[string]any{"xKey": xKey, "existingCollectionId": col.Id}), true
		}
	}
	return nil, false
}

// ownerProperties resolves the property definitions of a type or a
// collection by id — the shared definition surface — answering
// space.ErrNotFound when neither resolves.
func ownerProperties(ctx context.Context, sp space.Space, ownerId string) ([]space.PropertyDef, error) {
	if _, err := sp.Types().Get(ctx, ownerId); err == nil {
		return sp.Types().Properties(ctx, ownerId)
	}
	if _, err := sp.Collections().Get(ctx, ownerId); err != nil {
		return nil, err
	}
	return sp.Collections().Properties(ctx, ownerId)
}

func typeInfoToAPI(t space.TypeInfo) api.TypeInfo {
	// Builtin/registered types have clean literal ids (chat, editor, …) and no
	// caller-set xKey; report xKey=id so every type has a stable programmatic
	// handle (user types carry the xKey set at create / derived from name).
	xkey := t.XKey
	if xkey == "" && t.BuiltIn {
		xkey = t.Id
	}
	out := api.TypeInfo{
		Id:          t.Id,
		Name:        t.Name,
		Description: t.Description,
		IconCID:     t.IconCID,
		XKey:        xkey,
		BuiltIn:     t.BuiltIn,
		Hidden:      t.Hidden,
		Meta:        t.Meta,
	}
	if len(t.Layout) > 0 {
		if raw, err := json.Marshal(t.Layout); err == nil {
			out.Layout = raw
		}
	}
	return out
}

func propertyDefToAPI(p space.PropertyDef) api.PropertyDef {
	out := api.PropertyDef{
		Id:          p.Id,
		Name:        p.Name,
		Description: p.Description,
		XKey:        p.XKey,
		Kind:        propertyKindToString(p.Kind),
		Meta:        p.Meta,
		XFormat:     xformatToWire(p.XFormat),
	}
	if p.Scope != 0 {
		out.Scope = p.Scope.String()
	}
	if p.Items != nil {
		nested := propertyDefToAPI(*p.Items)
		out.Items = &nested
	}
	if len(p.Properties) > 0 {
		out.Properties = make([]api.PropertyDef, 0, len(p.Properties))
		for _, np := range p.Properties {
			out.Properties = append(out.Properties, propertyDefToAPI(np))
		}
	}
	if len(p.Required) > 0 {
		out.Required = append([]string(nil), p.Required...)
	}
	return out
}

func propertyKindToString(k space.PropertyKind) string {
	switch k {
	case space.PropertyKindString:
		return api.PropertyKindString
	case space.PropertyKindNumber:
		return api.PropertyKindNumber
	case space.PropertyKindBoolean:
		return api.PropertyKindBoolean
	case space.PropertyKindNull:
		return api.PropertyKindNull
	case space.PropertyKindArray:
		return api.PropertyKindArray
	case space.PropertyKindObject:
		return api.PropertyKindObject
	case space.PropertyKindDatetime:
		return api.PropertyKindDatetime
	default:
		return ""
	}
}

func propertyKindFromString(s string) (space.PropertyKind, bool) {
	switch s {
	case api.PropertyKindString:
		return space.PropertyKindString, true
	case api.PropertyKindNumber:
		return space.PropertyKindNumber, true
	case api.PropertyKindBoolean:
		return space.PropertyKindBoolean, true
	case api.PropertyKindNull:
		return space.PropertyKindNull, true
	case api.PropertyKindArray:
		return space.PropertyKindArray, true
	case api.PropertyKindObject:
		return space.PropertyKindObject, true
	case api.PropertyKindDatetime:
		return space.PropertyKindDatetime, true
	default:
		return 0, false
	}
}